most of the function working
Published on a branch, not entirely surehow git works so i dont want to
overwrite any potential progress. Includes a readTxtFile class that will
do all of the work. The tempmain class is what I used for testing and
will show how to create and use a readTxtFile object. The file I used
for testing is the input.txt file that I got from some random game on my
computer. I leave file reading and datapaths to callahan. Make fixes as
necessary but probably dont mess with the readTxtFile class.